oct-1396: split onboarding e2e
